#7 Fix: user 엔티티 객체를 사용하기 위해 수정

+++ src/main/java/chugpuff/chugpuff/service/LikeService.java
@@ -2,8 +2,10 @@ package chugpuff.chugpuff.service;
 
 import chugpuff.chugpuff.entity.Board;
 import chugpuff.chugpuff.entity.Like;
+import chugpuff.chugpuff.entity.User;
 import chugpuff.chugpuff.repository.BoardRepository;
 import chugpuff.chugpuff.repository.LikeRepository;
+import chugpuff.chugpuff.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,16 +20,20 @@ public class LikeService {
     @Autowired
     private BoardRepository boardRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     public int getLikesCount(int boardNo) {
         return likeRepository.countByBoard_BoardNo(boardNo); // 여기서 countByBoard_BoardNo를 호출합니다.
     }
 
     public Like addLike(int boardNo, String userId) {
         Optional<Board> board = boardRepository.findById(boardNo);
+        Optional<User> user = userRepository.findById(userId);
         if (board.isPresent()) {
             Like newLike = new Like();
             newLike.setBoard(board.get());
-            newLike.setUserId(userId);
+            newLike.setUser(user.get());
             return likeRepository.save(newLike);
         }
         return null;
@@ -35,7 +41,8 @@ public class LikeService {
 
     public void removeLike(int boardNo, String userId) {
         Optional<Board> board = boardRepository.findById(boardNo);
-        if (board.isPresent()) {
+        Optional<User> user = userRepository.findById(userId);
+        if (board.isPresent() && user.isPresent()) {
             Optional<Like> likeOptional = likeRepository.findByBoardAndUserId(board.get(), userId);
             if (likeOptional.isPresent()) {
                 likeRepository.delete(likeOptional.get());
@@ -45,14 +52,15 @@ public class LikeService {
 
     public void toggleLike(int boardNo, String userId) {
         Optional<Board> board = boardRepository.findById(boardNo);
-        if (board.isPresent()) {
+        Optional<User> user = userRepository.findById(userId);
+        if (board.isPresent() && user.isPresent()) {
             Optional<Like> likeOptional = likeRepository.findByBoardAndUserId(board.get(), userId);
             if (likeOptional.isPresent()) {
                 likeRepository.delete(likeOptional.get());
             } else {
                 Like newLike = new Like();
                 newLike.setBoard(board.get());
-                newLike.setUserId(userId);
+                newLike.setUser(user.get());
                 likeRepository.save(newLike);
             }
         }
